Add basic Facebook authentication
This isn't wired up to anything all that useful yet, but is actually
working which was my goal for the day.

The email address isn't coming back from Facebook yet, so I still need
to figure that out. I also need to figure out how to request the
permissions that we'll need for the app and where to store them, but
that can all wait for a future commit.

I expect that I'll want to move the passport code to its own file, and
the Queries class to its own file to make the code here more manageable,
but for now one big file does the trick.