Display currentUser's name in <NavBar>
When logged in, we want to show the current user's name in the nav bar.
This will eventually be a menu where they can change their settings and
log out and stuff like that.

Since I'm not very familiar with how this type of application is set up,
I decided to just issue an AJAX request to /api/me when the application
first loads to get the current user's info.