Replace connect-history-api-fallback dependency with simple fallback
I want to make my index.html a bit dynamic, so that I can inject the
app's initial state and I don't think this is currently possible since
it is built by the webpack plugin and served directly up the client.

I think I need to bring this into the app a little more so I can set up
a template engine, and the first step I see is by replacing the
connect-history-api-fallback dependency with the equivalent fallback
route for my express server.

Even if I don't end up needing this, I find this a lot more readable
than the dependency anyway, so I see this as a win.

Note that I didn't really test the production config so I'm not sure if
it actually works. My plan is to just wait to figure that out once I get
to the point of actually deploying this app somewhere.

Some of this setup was inspired by
https://github.com/christianalfoni/webpack-express-boilerplate