Use handlebars with express and webpack
I want to be able to bootstrap the initial state of the app onto the
page when it first loads, to make authentication a little nicer. To do
this, I need to make a templating engine, express, and webpack all play
together nicely. Thankfully, webpack gives us some stats about the files
we need to load that makes this possible.

Currently this almost works, with the exception of what seems like a
bug in webpack. More information at:

  https://github.com/webpack/webpack/issues/762#issuecomment-167166027