Make webpack, express, and handlebars actually work
Since I wasn't able to achieve consistent results by just using
webpack's Node API (see issue linked in previous commit), I decided to
work around it by using a webpack plugin that generates a stats.json
file, and then polling for it when starting up the server. This is
similar to a technique that I used previously in a different project
that used Rails.