Bootstrap app initial state with currentUser
This will make the app work better when it is first loaded with a logged
in user, since we won't have to wait for it to make an ajax request to
figure out who the current user is.

This all feels like a bit of a hack, so I'm not really sure if its a
good way to do this type of thing, but I'm rolling with it for now so I
can keep moving forward.